Allow different sources of configuration
All of the previous configuration logic has been moved to the
FileOrHashSource class. There are a few quirks to the original behavior,
so it was simpler to keep it a single source for now. This can likely be
simplified if we do not need to keep all of the original behavior.

For example, if the ability to load from a Hash is used solely for
testing, all of that logic could likely be removed, in favor of a config
source that only exists in the test code.